Add SBML, GML and network knowledge base files
The structures of the Gene Regulatory and the Protein signaling sub-networks that were implemented in CellCollective were exported in the standard  SBMLqual and GML formats as well as under the form of interaction matrices (1 denotes activator links and -1 and inhibitory links) to allow modelers to re-use the network in their favorite modeling platform. 

The knowledge base with literature support was also exported in a tabular format for both the GRN and the PPI layer.

Important note:  SBMLqual format was used to export/import the network structure (list of nodes, interactions and network layout). However, the computational model presented in this repository uses a semi-quantitative formalism, it is not a logical boolean model therefore the boolean rules that were automatically generated in the SBML file do not describe accurately the computational model. To run the computational model and reproduce the result of the paper (https://doi.org/10.1101/2021.09.27.461207), the equations, scripts and in-house algorithm presented in the current repository must be used.